ASSIGNED - bug 371006: Remove commands from UI thread https://bugs.eclipse.org/bugs/show_bug.cgi?id=371006

+++ org.eclipse.mylyn.reviews.r4e.ui/src/org/eclipse/mylyn/reviews/r4e/ui/internal/properties/general/AnomalyExtraProperties.java
@@ -142,9 +142,8 @@ public class AnomalyExtraProperties extends AnomalyBasicProperties {
 			return result;
 		}
 		if (ANOMALY_STATE_ID.equals(aId)) {
-			return R4EUIAnomalyExtended.getStates()[Integer.valueOf(((R4EUIAnomalyExtended) getElement()).getAnomaly()
-					.getState()
-					.getValue())];
+			return R4EUIAnomalyExtended.getStates()[Integer.valueOf(
+					((R4EUIAnomalyExtended) getElement()).getAnomaly().getState().getValue()).intValue()];
 		} else if (ANOMALY_NOT_ACCEPTED_REASON_ID.equals(aId)) {
 			if (null != ((R4EUIAnomalyExtended) getElement()).getAnomaly().getNotAcceptedReason()) {
 				return ((R4EUIAnomalyExtended) getElement()).getAnomaly().getNotAcceptedReason();
